Fixup linux installs

+++ install.sh
@@ -144,19 +144,28 @@ if [ ! -f "$HOME/.ssh/id_rsa.pub" ]; then
 fi
 
 if is_debian; then
-  fancy_echo "Installing Keybase ..." "$yellow"
-  curl --remote-name https://prerelease.keybase.io/keybase_amd64.deb
-  sudo dpkg -i keybase_amd64.deb
-  sudo apt-get install -f
-  run_keybase
 
-  fancy_echo "Installing Yarn ..." "$yellow"
-  curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
-  echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
+  if ! type keybase 2>/dev/null; then
+    fancy_echo "Installing Keybase ..." "$yellow"
+    curl --remote-name https://prerelease.keybase.io/keybase_amd64.deb
+    sudo dpkg -i keybase_amd64.deb
+    sudo apt-get install -f
+    run_keybase
+  fi
 
-  fancy_echo "Installing Google Chrome ..." "$yellow"
-  curl -sS https://dl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
-  echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" | sudo tee /etc/apt/sources.list.d/google-chrome.list
+  if [ ! -e /etc/apt/sources.list.d/yarn.list ]; then
+    fancy_echo "Installing Yarn ..." "$yellow"
+    curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
+    echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
+    sudo apt update
+    sudo apt install --no-install-recommends yarn
+  fi
+
+  if [ ! -e /etc/apt/sources.list.d/google-chrome.list ]; then
+    fancy_echo "Installing Google Chrome ..." "$yellow"
+    curl -sS https://dl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
+    echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" | sudo tee /etc/apt/sources.list.d/google-chrome.list
+  fi
 
   fancy_echo "Installing Google Drive" "$yellow"
   sudo add-apt-repository ppa:alessandro-strada/ppa
@@ -167,14 +176,15 @@ if is_debian; then
   fancy_echo "Installing Papirus GTK Icons ..." "$yellow"
   sudo add-apt-repository ppa:papirus/papirus
 
-  fancy_echo "Installing PostgreSQL 9.6..." "$yellow"
-  echo "deb http://apt.postgresql.org/pub/repos/apt/ bionic-pgdg main" | sudo tee /etc/apt/sources.list.d/pgdg.list
-  wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
+  if [ ! -e /etc/apt/sources.list.d/pgdg.list ]; then
+    fancy_echo "Installing PostgreSQL 9.6..." "$yellow"
+    echo "deb http://apt.postgresql.org/pub/repos/apt/ bionic-pgdg main" | sudo tee /etc/apt/sources.list.d/pgdg.list
+    wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
+  fi
 
   sudo apt update
 
-  sudo apt install postgresql-9.6 google-chrome-stable papirus-icon-theme
-  adapta-gtk-theme google-drive-ocamlfuse
+  sudo apt install postgresql-9.6 google-chrome-stable papirus-icon-theme adapta-gtk-theme google-drive-ocamlfuse
 
   pip3 install pywal
 fi
@@ -261,14 +271,19 @@ ln -sf "$HOME/dotfiles/prompt_bernheisel_setup" "$HOME/.zprezto/modules/prompt/f
 #### Brew installs
 column
 fancy_echo "Installing programs" "$yellow"
-if brew list | grep -Fq brew-cask; then
-  fancy_echo "Uninstalling old Homebrew-Cask ..." "$yellow"
-  brew uninstall --force brew-cask
-fi
 brew update
 brew bundle check || brew bundle
 brew cleanup
-brew prune
+
+if is_linux; then
+  brew unlink postgresql
+  brew unlink python
+  brew unlink python2
+  brew unlink python3
+  brew unlink rsync
+  brew unlink xdpyinfo
+  brew unlink systemd
+fi
 
 
 #### Setup dotfiles
@@ -309,6 +324,25 @@ column
 fancy_echo "Symlinking config files" "$yellow"
 for f in "${files[@]}"; do
   ln -fs "$folder/$f" "$HOME/.$f"
+done
+
+configfiles=(
+  bspwm
+  compton
+  gtk-3.0
+  kitty
+  nvim
+  polybar
+  ranger
+  rofi
+  sxkhd
+  wtf
+  gtkrc
+  gtkrc-2.0
+)
+
+for f in "${configfiles[@]}"; do
+  ln -fs "$folder/config/$f" "$HOME/.config/$f"
 done
 
 if [ ! -e "$HOME/.secrets" ]; then
@@ -444,7 +478,7 @@ if type asdf &> /dev/null; then
 fi
 
 ln -fs "$HOME/.config" "$HOME/dotfiles/.config"
-ln -fs "$HOME/dotfiles/ctags.d" "$HOME/.ctags.d"
+ln -fs "$HOME/.ctags.d" "$HOME/dotfiles/ctags.d"
 
 fancy_echo "Installing language servers" "$yellow"
 yarn_install_or_update vscode-json-languageserver-bin
